+++ packages/expo-router/CHANGELOG.md
@@ -64,6 +64,7 @@
 
 ### 💡 Others
 
+- Migrate Slot, JavaScript Stack, and ExperimentalStack to the `standard-navigation` integration.
 - Remove the ignored `linking.enabled` option. ([#49103](https://github.com/expo/expo/pull/49103) by [@Ubax](https://github.com/Ubax))
 - Render only the focused tab route during the first render. ([#48618](https://github.com/expo/expo/pull/48618) by [@Ubax](https://github.com/Ubax))
 - Order JS Tabs, NativeTabs, headless tabs and Drawer by `routeNames` order ([#48374](https://github.com/expo/expo/pull/48374) by [@Ubax](https://github.com/Ubax))